Adopt new engine.

+++ src/ru/nsu/ccfit/zuev/osu/game/cursor/flashlight/MainFlashLightSprite.java
@@ -1,15 +1,15 @@
 package ru.nsu.ccfit.zuev.osu.game.cursor.flashlight;
 
-import org.anddev.andengine.entity.modifier.ScaleModifier;
-import org.anddev.andengine.opengl.texture.region.TextureRegion;
+import org.andengine.entity.modifier.ScaleModifier;
+import org.andengine.opengl.texture.region.TextureRegion;
 
 import ru.nsu.ccfit.zuev.osu.ResourceManager;
 
 
 public class MainFlashLightSprite extends FlashlightAreaSizedSprite {
     private static final TextureRegion DEFAULT_TEXTURE = ResourceManager.getInstance().getTexture("flashlight_cursor");
-    public static final int TEXTURE_WIDTH = DEFAULT_TEXTURE.getWidth();
-    public static final int TEXTURE_HEIGHT = DEFAULT_TEXTURE.getHeight();
+    public static final float TEXTURE_WIDTH = DEFAULT_TEXTURE.getWidth();
+    public static final float TEXTURE_HEIGHT = DEFAULT_TEXTURE.getHeight();
     public final float AREA_CHANGE_FADE_DURATION = 0.8f;
     public float currentSize = BASE_SIZE;
